--- a/clovers-app/README.md
+++ b/clovers-app/README.md
@@ -1,5 +1,0 @@
-# clovers 🍀 ray tracing in rust 🦀
-
-This directory contains an example GUI application utilizing `clovers` as the rendering library.
-
-NOTE: this is currently broken and needs a complete rewrite.