--- a/clovers/src/objects.rs
+++ b/clovers/src/objects.rs
@@ -54,7 +54,7 @@ pub enum Object {
     Sphere(SphereInit),
     #[cfg(feature = "stl")]
     /// STL object initializer
-    STL(STL),
+    STL(STLInit),
     /// Translate object initializer
     Translate(TranslateInit),
     /// Triangle object initializer
@@ -89,8 +89,9 @@ impl From<Object> for Hitable {
             }
             Object::Sphere(x) => Hitable::Sphere(Sphere::new(x.center, x.radius, x.material)),
             #[cfg(feature = "stl")]
-            Object::STL(_x) => {
-                unimplemented!("Do not call .into() directly, see objects_to_flat_hitablelist");
+            Object::STL(x) => {
+                // TODO: time
+                Hitable::STL(STL::new(x, 0.0, 1.0))
             }
             Object::Translate(x) => {
                 let obj = *x.object;