--- a/clovers/src/objects.rs
+++ b/clovers/src/objects.rs
@@ -43,7 +43,7 @@ pub enum Object {
     /// FlipFace object initializer
     FlipFace(FlipFaceInit),
     /// MovingSphere object initializer
-    MovingSphere(MovingSphere),
+    MovingSphere(MovingSphereInit),
     /// ObjectList object initializer
     ObjectList(ObjectList),
     /// Quad object initializer
@@ -76,7 +76,8 @@ impl From<Object> for Hitable {
                 Hitable::FlipFace(FlipFace::new(obj))
             }
             Object::MovingSphere(x) => Hitable::MovingSphere(MovingSphere::new(
-                x.center_0, x.center_1, x.time_0, x.time_1, x.radius, x.material,
+                // TODO: time
+                x.center_0, x.center_1, 0.0, 1.0, x.radius, x.material,
             )),
             Object::ObjectList(_x) => {
                 unimplemented!("Do not call .into() directly, see objects_to_flat_hitablelist");