--- a/clovers/src/objects/moving_sphere.rs
+++ b/clovers/src/objects/moving_sphere.rs
@@ -3,6 +3,21 @@
 use crate::{aabb::AABB, hitable::HitRecord, materials::Material, ray::Ray, Float, Vec3, PI};
 use rand::rngs::SmallRng;
 
+#[derive(Clone, Copy, Debug)]
+#[cfg_attr(feature = "serde-derive", derive(serde::Serialize, serde::Deserialize))]
+/// `SphereInit` structure describes the necessary data for constructing a [Sphere]. Used with [serde] when importing [`SceneFile`](crate::scenes::SceneFile)s.
+pub struct MovingSphereInit {
+    /// Center point of the sphere at time_0
+    pub center_0: Vec3,
+    /// Center point of the sphere at time_1
+    pub center_1: Vec3,
+    /// Radius of the sphere.
+    pub radius: Float,
+    #[cfg_attr(feature = "serde-derive", serde(default))]
+    /// Material of the sphere.
+    pub material: Material,
+}
+
 #[derive(Debug, Clone)]
 #[cfg_attr(feature = "serde-derive", derive(serde::Serialize, serde::Deserialize))]
 /// A moving sphere object. This is represented by one `radius`, two center points `center_0` `center_1`, two times `time_0` `time_1`, and a [Material]. Any [Rays](Ray) hitting the object will also have an internal `time` value, which will be used for determining the interpolated position of the sphere at that time. With lots of rays hitting every pixel but at randomized times, we get temporal multiplexing and an approximation of perceived motion blur.