--- a/scenes/the_next_week.json
+++ b/scenes/the_next_week.json
@@ -11006,11 +11006,6 @@
         "radius": 80,
         "material": {
           "Lambertian": {
-            "albedo": {
-              "NoiseTexture": {
-                "scale": 0.1
-              }
-            }
           }
         }
       }