- Deprecate / delete GUI version
- BVHNode `count()` utility method
- STLInit and dedicated BVHNode trees per STL object
- MovingSphereInit, forgot to add earlier
  
Before:
```
Benchmark 1: ./target/release/clovers-cli -w 512 -h 256 -s 100 -q -i scenes/scene.json -o temp.png
  Time (mean ± σ):      3.796 s ±  0.065 s    [User: 202.846 s, System: 0.466 s]
  Range (min … max):    3.744 s …  3.959 s    10 runs
```

After:
```
Benchmark 1: ./target/release/clovers-cli -w 512 -h 256 -s 100 -q -i scenes/scene.json -o temp.png
  Time (mean ± σ):      2.893 s ±  0.026 s    [User: 141.226 s, System: 0.488 s]
  Range (min … max):    2.864 s …  2.931 s    10 runs
```

Approximately 25% decrease in rendering time for some test scenes.